Tests/LibWeb/Text/expected/async-scrolling/orthogonal-overflow-padding-wheel-target.txt

client width: 100
scroll width: 100
horizontal wheel target: none
